requirements.txt
pyenchant
django-filebrowser-no-grappelli
Pillow
coverage
